Port to windows

+++ C/src/light_vm/lightvm.h
@@ -252,7 +252,7 @@ Light_VM_Instruction_Info light_vm_push_fmt(Light_VM_State* vm_state, const char
 Light_VM_Instruction      light_vm_instruction_get(const char* s, uint64_t* immediate);
 void*                     light_vm_push_data_segment(Light_VM_State* vm_state, Light_VM_Data data);
 void*                     light_vm_push_bytes_data_segment(Light_VM_State* vm_state, uint8_t* bytes, int32_t byte_count);
-int32_t                   light_vm_patch_immediate_distance(Light_VM_Instruction_Info from, Light_VM_Instruction_Info to);
+int64_t                   light_vm_patch_immediate_distance(Light_VM_Instruction_Info from, Light_VM_Instruction_Info to);
 uint64_t                  light_vm_offset_from_current_instruction(Light_VM_State* state, Light_VM_Instruction_Info from);
 uint8_t                   light_vm_patch_to_current_instruction(Light_VM_State* state, Light_VM_Instruction_Info to);
 void                      light_vm_patch_from_to_current_instruction(Light_VM_State* state, Light_VM_Instruction_Info from);
